+++ sahi/__init__.py
@@ -1,3 +1,6 @@
-__version__ = "0.10.7"
+__version__ = "0.11.6"
 
+from sahi.annotation import BoundingBox, Category, Mask
 from sahi.auto_model import AutoDetectionModel
+from sahi.models.base import DetectionModel
+from sahi.prediction import ObjectPrediction
