+++ sahi/models/__init__.py
@@ -1,0 +1,1 @@
+from . import base, detectron2, huggingface, mmdet, torchvision, yolov5
