+++ tests/test_detectron2.py
@@ -3,7 +3,7 @@
 
 import unittest
 
-from sahi.model import Detectron2DetectionModel
+from sahi.models.detectron2 import Detectron2DetectionModel
 from sahi.utils.cv import read_image
 from sahi.utils.detectron2 import Detectron2TestConstants
 from sahi.utils.import_utils import get_package_info
@@ -14,7 +14,8 @@ IMAGE_SIZE = 320
 
 # note that detectron2 binaries are available only for linux
 
-if get_package_info("torch", verbose=False)[1] == "1.10.2":
+torch_version = get_package_info("torch", verbose=False)[1]
+if "1.10." in torch_version:
 
     class TestDetectron2DetectionModel(unittest.TestCase):
         def test_load_model(self):
@@ -86,7 +87,7 @@ if get_package_info("torch", verbose=False)[1] == "1.10.2":
             self.assertEqual(len(object_prediction_list), 16)
             self.assertEqual(object_prediction_list[0].category.id, 2)
             self.assertEqual(object_prediction_list[0].category.name, "car")
-            predicted_bbox = object_prediction_list[0].bbox.to_coco_bbox()
+            predicted_bbox = object_prediction_list[0].bbox.to_xywh()
             desired_bbox = [831, 303, 42, 43]
             margin = 3
             for ind, point in enumerate(predicted_bbox):
@@ -95,7 +96,7 @@ if get_package_info("torch", verbose=False)[1] == "1.10.2":
 
             self.assertEqual(object_prediction_list[5].category.id, 2)
             self.assertEqual(object_prediction_list[5].category.name, "car")
-            predicted_bbox = object_prediction_list[2].bbox.to_coco_bbox()
+            predicted_bbox = object_prediction_list[2].bbox.to_xywh()
             desired_bbox = [383, 277, 36, 29]
             margin = 3
             for ind, point in enumerate(predicted_bbox):
@@ -128,7 +129,7 @@ if get_package_info("torch", verbose=False)[1] == "1.10.2":
             self.assertEqual(len(object_prediction_list), 13)
             self.assertEqual(object_prediction_list[0].category.id, 2)
             self.assertEqual(object_prediction_list[0].category.name, "car")
-            predicted_bbox = object_prediction_list[0].bbox.to_coco_bbox()
+            predicted_bbox = object_prediction_list[0].bbox.to_xywh()
             desired_bbox = [321, 324, 59, 38]
             margin = 3
             for ind, point in enumerate(predicted_bbox):
@@ -137,7 +138,7 @@ if get_package_info("torch", verbose=False)[1] == "1.10.2":
 
             self.assertEqual(object_prediction_list[5].category.id, 2)
             self.assertEqual(object_prediction_list[5].category.name, "car")
-            predicted_bbox = object_prediction_list[5].bbox.to_coco_bbox()
+            predicted_bbox = object_prediction_list[5].bbox.to_xywh()
             desired_bbox = [719, 243, 27, 30]
             margin = 3
             for ind, point in enumerate(predicted_bbox):
@@ -145,7 +146,7 @@ if get_package_info("torch", verbose=False)[1] == "1.10.2":
                     raise AssertionError(f"desired_bbox: {desired_bbox}, predicted_bbox: {predicted_bbox}")
 
         def test_get_prediction_detectron2(self):
-            from sahi.model import Detectron2DetectionModel
+            from sahi.models.detectron2 import Detectron2DetectionModel
             from sahi.predict import get_prediction
             from sahi.utils.detectron2 import Detectron2TestConstants
 
@@ -194,7 +195,7 @@ if get_package_info("torch", verbose=False)[1] == "1.10.2":
             self.assertEqual(num_car, 16)
 
         def test_get_sliced_prediction_detectron2(self):
-            from sahi.model import Detectron2DetectionModel
+            from sahi.models.detectron2 import Detectron2DetectionModel
             from sahi.predict import get_sliced_prediction
             from sahi.utils.detectron2 import Detectron2TestConstants
 
@@ -239,7 +240,7 @@ if get_package_info("torch", verbose=False)[1] == "1.10.2":
             object_prediction_list = prediction_result.object_prediction_list
 
             # compare
-            self.assertEqual(len(object_prediction_list), 18)
+            self.assertEqual(len(object_prediction_list), 19)
             num_person = 0
             for object_prediction in object_prediction_list:
                 if object_prediction.category.name == "person":
@@ -254,7 +255,7 @@ if get_package_info("torch", verbose=False)[1] == "1.10.2":
             for object_prediction in object_prediction_list:
                 if object_prediction.category.name == "car":
                     num_car += 1
-            self.assertEqual(num_car, 18)
+            self.assertEqual(num_car, 19)
 
 
 if __name__ == "__main__":
