+++ tests/test_highlevelapi.py
@@ -1,12 +1,12 @@
 # OBSS SAHI Tool
-# Code written by Fatih C Akyon, 2020.
+# Code written by Fatih C Akyon, 2022.
 
 import unittest
 
 
-class TestAnnotation(unittest.TestCase):
+class TestHighLevelApi(unittest.TestCase):
     def test_bounding_box(self):
-        from sahi.annotation import BoundingBox
+        from sahi import BoundingBox
 
         bbox_minmax = [30, 30, 100, 150]
         shift_amount = [50, 40]
@@ -23,7 +23,7 @@ class TestAnnotation(unittest.TestCase):
         self.assertEqual(shifted_bbox.to_xyxy(), [80, 70, 150, 190])
 
     def test_category(self):
-        from sahi.annotation import Category
+        from sahi import Category
 
         category_id = 1
         category_name = "car"
@@ -32,7 +32,7 @@ class TestAnnotation(unittest.TestCase):
         self.assertEqual(category.name, category_name)
 
     def test_mask(self):
-        from sahi.annotation import Mask
+        from sahi import Mask
 
         coco_segmentation = [[1, 1, 325, 125, 250, 200, 5, 200]]
         full_shape_height, full_shape_width = 500, 600
@@ -44,8 +44,8 @@ class TestAnnotation(unittest.TestCase):
         self.assertEqual(mask.full_shape_width, full_shape_width)
         self.assertEqual(mask.bool_mask[11, 2], True)
 
-    def test_object_annotation(self):
-        from sahi.annotation import ObjectAnnotation
+    def test_object_prediction(self):
+        from sahi import ObjectPrediction
 
         bbox = [100, 200, 150, 230]
         coco_bbox = [bbox[0], bbox[1], bbox[2] - bbox[0], bbox[3] - bbox[1]]
@@ -56,7 +56,7 @@ class TestAnnotation(unittest.TestCase):
         image_width = 1920
         full_shape = [image_height, image_width]
 
-        object_annotation1 = ObjectAnnotation(
+        object_annotation1 = ObjectPrediction(
             bbox=bbox,
             category_id=category_id,
             category_name=category_name,
@@ -64,14 +64,14 @@ class TestAnnotation(unittest.TestCase):
             full_shape=full_shape,
         )
 
-        object_annotation2 = ObjectAnnotation.from_coco_annotation_dict(
+        object_annotation2 = ObjectPrediction.from_coco_annotation_dict(
             annotation_dict={"bbox": coco_bbox, "category_id": category_id, "segmentation": []},
             category_name=category_name,
             full_shape=full_shape,
             shift_amount=shift_amount,
         )
 
-        object_annotation3 = ObjectAnnotation.from_coco_bbox(
+        object_annotation3 = ObjectPrediction.from_coco_bbox(
             bbox=coco_bbox,
             category_id=category_id,
             category_name=category_name,
@@ -99,6 +99,15 @@ class TestAnnotation(unittest.TestCase):
         self.assertEqual(object_annotation3.bbox.maxy, bbox[3])
         self.assertEqual(object_annotation3.category.id, category_id)
         self.assertEqual(object_annotation3.category.name, category_name)
+
+    def test_detection_model(self):
+        from sahi import DetectionModel
+
+        MODEL_PATH = "model_path"
+        IAMGE_SIZE = 640
+        detection_model = DetectionModel(model_path="model_path", image_size=IAMGE_SIZE, load_at_init=False)
+        self.assertEqual(detection_model.model_path, MODEL_PATH)
+        self.assertEqual(detection_model.image_size, IAMGE_SIZE)
 
 
 if __name__ == "__main__":
