+++ tests/test_predict.py
@@ -27,7 +27,7 @@ class TestPredict(unittest.TestCase):
         from sahi.prediction import ObjectPrediction
 
     def test_get_prediction_mmdet(self):
-        from sahi.model import MmdetDetectionModel
+        from sahi.models.mmdet import MmdetDetectionModel
         from sahi.predict import get_prediction
         from sahi.utils.mmdet import MmdetTestConstants, download_mmdet_yolox_tiny_model
 
@@ -73,7 +73,7 @@ class TestPredict(unittest.TestCase):
         self.assertEqual(num_car, 2)
 
     def test_get_prediction_yolov5(self):
-        from sahi.model import Yolov5DetectionModel
+        from sahi.models.yolov5 import Yolov5DetectionModel
         from sahi.predict import get_prediction
         from sahi.utils.yolov5 import Yolov5TestConstants, download_yolov5n_model
 
@@ -166,7 +166,7 @@ class TestPredict(unittest.TestCase):
         self.assertEqual(num_car, 2)
 
     def test_get_sliced_prediction_mmdet(self):
-        from sahi.model import MmdetDetectionModel
+        from sahi.models.mmdet import MmdetDetectionModel
         from sahi.predict import get_sliced_prediction
         from sahi.utils.mmdet import MmdetTestConstants, download_mmdet_yolox_tiny_model
 
@@ -213,7 +213,7 @@ class TestPredict(unittest.TestCase):
         object_prediction_list = prediction_result.object_prediction_list
 
         # compare
-        self.assertEqual(len(object_prediction_list), 14)
+        self.assertEqual(len(object_prediction_list), 15)
         num_person = 0
         for object_prediction in object_prediction_list:
             if object_prediction.category.name == "person":
@@ -228,10 +228,10 @@ class TestPredict(unittest.TestCase):
         for object_prediction in object_prediction_list:
             if object_prediction.category.name == "car":
                 num_car += 1
-        self.assertEqual(num_car, 14)
+        self.assertEqual(num_car, 15)
 
     def test_get_sliced_prediction_yolov5(self):
-        from sahi.model import Yolov5DetectionModel
+        from sahi.models.yolov5 import Yolov5DetectionModel
         from sahi.predict import get_sliced_prediction
         from sahi.utils.yolov5 import Yolov5TestConstants, download_yolov5n_model
 
@@ -277,7 +277,7 @@ class TestPredict(unittest.TestCase):
         object_prediction_list = prediction_result.object_prediction_list
 
         # compare
-        self.assertEqual(len(object_prediction_list), 10)
+        self.assertEqual(len(object_prediction_list), 11)
         num_person = 0
         for object_prediction in object_prediction_list:
             if object_prediction.category.name == "person":
@@ -292,7 +292,7 @@ class TestPredict(unittest.TestCase):
         for object_prediction in object_prediction_list:
             if object_prediction.category.name == "car":
                 num_car += 1
-        self.assertEqual(num_car, 10)
+        self.assertEqual(num_car, 11)
 
     def test_coco_json_prediction(self):
         from sahi.predict import predict
