+++ tests/test_shapelyutils.py
@@ -28,7 +28,7 @@ class TestShapelyUtils(unittest.TestCase):
 
     def test_shapely_annotation(self):
         # init shapely_annotation from coco segmentation
-        segmentation = [[1, 1, 325, 125, 250, 200, 5, 200]]
+        segmentation = [[1, 1, 325, 125.2, 250, 200, 5, 200]]
         shapely_multipolygon = get_shapely_multipolygon(segmentation)
         shapely_annotation = ShapelyAnnotation.from_coco_segmentation(segmentation)
 
@@ -51,12 +51,12 @@ class TestShapelyUtils(unittest.TestCase):
                 ]
             ],
         )
-        coco_bbox = shapely_annotation.to_coco_bbox()
+        coco_bbox = shapely_annotation.to_xywh()
         self.assertEqual(
             coco_bbox,
             [1, 1, 324, 199],
         )
-        voc_bbox = shapely_annotation.to_voc_bbox()
+        voc_bbox = shapely_annotation.to_xyxy()
         self.assertEqual(
             voc_bbox,
             [1, 1, 325, 200],
@@ -96,12 +96,12 @@ class TestShapelyUtils(unittest.TestCase):
                 ]
             ],
         )
-        coco_bbox = shapely_annotation.to_coco_bbox()
+        coco_bbox = shapely_annotation.to_xywh()
         self.assertEqual(
             coco_bbox,
             [1, 1, 100, 100],
         )
-        voc_bbox = shapely_annotation.to_voc_bbox()
+        voc_bbox = shapely_annotation.to_xyxy()
         self.assertEqual(
             voc_bbox,
             [1, 1, 101, 101],
@@ -150,9 +150,9 @@ class TestShapelyUtils(unittest.TestCase):
             ],
         )
 
-        self.assertEqual(intersection_shapely_annotation.to_coco_bbox(), [0, 0, 256, 199])
+        self.assertEqual(intersection_shapely_annotation.to_xywh(), [0, 0, 256, 199])
 
-        self.assertEqual(intersection_shapely_annotation.to_voc_bbox(), [0, 0, 256, 199])
+        self.assertEqual(intersection_shapely_annotation.to_xyxy(), [0, 0, 256, 199])
 
     def test_get_empty_intersection(self):
         x, y, width, height = 300, 300, 256, 256
@@ -165,7 +165,7 @@ class TestShapelyUtils(unittest.TestCase):
 
         self.assertEqual(intersection_shapely_annotation.area, 0)
 
-        self.assertEqual(intersection_shapely_annotation.to_coco_bbox(), [])
+        self.assertEqual(intersection_shapely_annotation.to_xywh(), [])
 
 
 if __name__ == "__main__":
