+++ tests/test_torchvision.py
@@ -4,6 +4,8 @@
 
 import unittest
 
+import numpy as np
+
 from sahi.utils.cv import read_image
 from sahi.utils.torchvision import TorchVisionTestConstants
 
@@ -16,7 +18,7 @@ class TestTorchVisionDetectionModel(unittest.TestCase):
     def test_load_model(self):
         from torchvision.models.detection.faster_rcnn import RoIHeads
 
-        from sahi.model import TorchVisionDetectionModel
+        from sahi.models.torchvision import TorchVisionDetectionModel
 
         torchvision_detection_model = TorchVisionDetectionModel(
             config_path=TorchVisionTestConstants.FASTERRCNN_CONFIG_PATH,
@@ -30,7 +32,7 @@ class TestTorchVisionDetectionModel(unittest.TestCase):
     def test_load_model_without_config_path(self):
         from torchvision.models.detection.faster_rcnn import RoIHeads
 
-        from sahi.model import TorchVisionDetectionModel
+        from sahi.models.torchvision import TorchVisionDetectionModel
 
         torchvision_detection_model = TorchVisionDetectionModel(
             confidence_threshold=CONFIDENCE_THRESHOLD,
@@ -44,7 +46,7 @@ class TestTorchVisionDetectionModel(unittest.TestCase):
         import torchvision
         from torchvision.models.detection.ssd import SSDHead
 
-        from sahi.model import TorchVisionDetectionModel
+        from sahi.models.torchvision import TorchVisionDetectionModel
 
         NUM_CLASSES = 15
         PRETRAINED = False
@@ -61,7 +63,7 @@ class TestTorchVisionDetectionModel(unittest.TestCase):
         self.assertEqual(isinstance(torchvision_detection_model.model.head, SSDHead), True)
 
     def test_perform_inference_without_mask_output(self):
-        from sahi.model import TorchVisionDetectionModel
+        from sahi.models.torchvision import TorchVisionDetectionModel
 
         # init model
         torchvision_detection_model = TorchVisionDetectionModel(
@@ -110,7 +112,7 @@ class TestTorchVisionDetectionModel(unittest.TestCase):
             self.assertTrue(scores[score_ind] >= 0)
 
     def test_convert_original_predictions_without_mask_output(self):
-        from sahi.model import TorchVisionDetectionModel
+        from sahi.models.torchvision import TorchVisionDetectionModel
 
         torchvision_detection_model = TorchVisionDetectionModel(
             config_path=TorchVisionTestConstants.FASTERRCNN_CONFIG_PATH,
@@ -136,10 +138,12 @@ class TestTorchVisionDetectionModel(unittest.TestCase):
         self.assertEqual(len(object_prediction_list), 7)
         self.assertEqual(object_prediction_list[0].category.id, 3)
         self.assertEqual(object_prediction_list[0].category.name, "car")
-        self.assertEqual(object_prediction_list[0].bbox.to_coco_bbox(), [315, 309, 65, 57])
+        np.testing.assert_almost_equal(
+            object_prediction_list[0].bbox.to_xywh(), [315.79, 309.33, 64.28, 56.94], decimal=1
+        )
 
     def test_convert_original_predictions_with_mask_output(self):
-        from sahi.model import TorchVisionDetectionModel
+        from sahi.models.torchvision import TorchVisionDetectionModel
 
         torchvision_detection_model = TorchVisionDetectionModel(
             config_path=TorchVisionTestConstants.FASTERRCNN_CONFIG_PATH,
@@ -165,10 +169,12 @@ class TestTorchVisionDetectionModel(unittest.TestCase):
         self.assertEqual(len(object_prediction_list), 7)
         self.assertEqual(object_prediction_list[0].category.id, 3)
         self.assertEqual(object_prediction_list[0].category.name, "car")
-        self.assertEqual(object_prediction_list[0].bbox.to_coco_bbox(), [315, 309, 65, 57])
+        np.testing.assert_almost_equal(
+            object_prediction_list[0].bbox.to_xywh(), [315.79, 309.33, 64.28, 56.94], decimal=1
+        )
 
     def test_get_prediction_torchvision(self):
-        from sahi.model import TorchVisionDetectionModel
+        from sahi.models.torchvision import TorchVisionDetectionModel
         from sahi.predict import get_prediction
 
         # init model
@@ -200,10 +206,12 @@ class TestTorchVisionDetectionModel(unittest.TestCase):
         self.assertEqual(len(object_prediction_list), 7)
         self.assertEqual(object_prediction_list[0].category.id, 3)
         self.assertEqual(object_prediction_list[0].category.name, "car")
-        self.assertEqual(object_prediction_list[0].bbox.to_coco_bbox(), [315, 309, 65, 57])
+        np.testing.assert_almost_equal(
+            object_prediction_list[0].bbox.to_xywh(), [315.79, 309.33, 64.28, 56.94], decimal=1
+        )
 
     def test_get_sliced_prediction_torchvision(self):
-        from sahi.model import TorchVisionDetectionModel
+        from sahi.models.torchvision import TorchVisionDetectionModel
         from sahi.predict import get_sliced_prediction
 
         # init model
@@ -246,10 +254,12 @@ class TestTorchVisionDetectionModel(unittest.TestCase):
         object_prediction_list = prediction_result.object_prediction_list
 
         # compare
-        self.assertEqual(len(object_prediction_list), 18)
+        self.assertEqual(len(object_prediction_list), 20)
         self.assertEqual(object_prediction_list[0].category.id, 3)
         self.assertEqual(object_prediction_list[0].category.name, "car")
-        self.assertEqual(object_prediction_list[0].bbox.to_coco_bbox(), [765, 259, 29, 25])
+        np.testing.assert_almost_equal(
+            object_prediction_list[0].bbox.to_xywh(), [765.81, 259.37, 28.62, 24.63], decimal=1
+        )
 
 
 if __name__ == "__main__":
